including validation in method getHogarMascota

+++ src/main/java/personas/Posicion.java
@@ -1,17 +1,17 @@
 package personas;
 
 public class Posicion {
-    private int longitud;
-    private int latitud;
+    private double longitud;
+    private double latitud;
 
-  public Posicion(int longitud, int latitud) {
+  public Posicion(double longitud, double latitud) {
     this.longitud = longitud;
     this.latitud = latitud;
   }
 
   public Integer distanciaA(Posicion posicion) {
-    int resultadoLongitud = this.longitud - posicion.longitud;
-    int resultadoLatitud = this.latitud - posicion.latitud;
+    double resultadoLongitud = this.longitud - posicion.longitud;
+    double resultadoLatitud = this.latitud - posicion.latitud;
     return (int) Math.sqrt(Math.pow(resultadoLatitud, 2) + Math.pow(resultadoLongitud, 2));
   }
 }
